refactor(minor): 💡 コンポーネントを整理
