+++ src/components/talks/EventWrapper/TalkTypeLabel/index.tsx
